Removendo utils.h

+++ src/game/enemy.c
@@ -5,7 +5,7 @@ static void _free(Enemy* this){
     free(this);
 }
 
-Enemy* new_Enemy(Int64 x, Int64 y){
+Enemy* new_Enemy(int x, int y){
     Enemy* this = malloc(sizeof(Enemy));
     this->x = x;
     this->y = y;
